feat(neomutt)!: configuration structure changes
These changes are mostly aimed at increasing credential access speed, as
well as making it so as few credentials are read by the program at a
time.

Startup- and shutdown-hooks are defined to decrypt the GPG-encrypted
credential file and to remove it when the program quits (yes, the file
is not removed if NeoMutt is killed, that'd require another watcher
script which just pushes the problem down the line).

Runtime hooks are divided into "generic" for settings all accounts share,
and specifics (defined by the $my_host variable) for each e-mail host.
- Account-hooks only load credentials necessary to log into accounts.
- Folder-hooks set cache directories, from (for send-hook) and set where
  necessary IMAP directories are.
- Send-hook loads SMTP credentials for sending e-mails and sets the sent
  e-mail IMAP directory.